Stroke_rehab/requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
brainflow>=5.10.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
pyserial>=3.5
